codebase/UserServices/pyproject.toml
[tool.poetry]
name = "userservices"
version = "0.1.0"
description = ""
authors = ["NxtGen-Dev-ZAH <zaheer.ahmad.contact@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
sqlmodel = "^0.0.21"
fastapi = "^0.111.1"
uvicorn = "^0.30.3"
aiokafka = "^0.11.0"
protobuf = "^5.27.2"
psycopg = "^3.2.1"
pydantic = "^2.8.2"
python-jose = "^3.3.0"
bcrypt = "3.2.0"
pytest = "^8.3.3"
pytest-asyncio = "^0.24.0"
httpx = "^0.27.2"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
